Split test_links to run twice, based on env-var
In the "run against all notebooks" mode, we test all the links, but only check links which are actually files in this repo, so the test is quick.
In the "run against changed notebooks" mode, we test all the links.

Thus, a PR that changes one notebook should no longer fail due to a broken link in another notebook